Make sure volume are returned as an int

+++ mopidy/utils/__init__.py
@@ -24,7 +24,7 @@ def rescale(v, old=None, new=None):
     new_min, new_max = new
     old_min, old_max = old
     scaling = float(new_max - new_min) / (old_max - old_min)
-    return round(scaling * (v - old_min) + new_min)
+    return int(round(scaling * (v - old_min) + new_min))
 
 
 def import_module(name):
